+++ 19_Mini-Capstone/Capstone/Classes/CateringItemsSubclasses/Appetizers.cs
@@ -6,7 +6,7 @@ namespace Capstone.Classes.CateringItemsSubclasses
 {
     class Appetizers : CateringItem
     {
-        public Appetizers(string name, double price, string productId) : base(name, price, productId)
+        public Appetizers(string name, decimal price, string productId) : base(name, price, productId)
         {
 
         }
